refactor: separate core Observable behavior from State
