+++ packages/state/src/state.ts
@@ -4,7 +4,9 @@ import {
   event,
   Observable,
   Notify,
-  PENDING_KEYS
+  PENDING_KEYS,
+  observing,
+  observe
 } from './observable';
 
 const define = Object.defineProperty;
@@ -148,26 +150,7 @@ abstract class State implements Observable {
   }
 
   [Observable](callback: Observable.Callback, required?: boolean) {
-    const watch = new Set<unknown>();
-    const proxy = Object.create(this);
-
-    addListener(this, (key) => {
-      if (watch.has(key)) callback();
-    });
-
-    OBSERVER.set(proxy, (key, value) => {
-      if (value === undefined && required)
-        throw new Error(`${this}.${key} is required in this context.`);
-
-      watch.add(key);
-
-      if (value instanceof Object && Observable in value)
-        return value[Observable](callback, required) || value;
-
-      return value;
-    });
-
-    return proxy as this;
+    return observe(this, callback, required);
   }
 
   /**
@@ -579,7 +562,7 @@ function manage(
   const state = STATE.get(target)!;
 
   function get(this: State) {
-    return follow(this, key, state[key]);
+    return observing(this, key, state[key]);
   }
 
   function set(value: unknown, silent?: boolean) {
@@ -636,16 +619,6 @@ function values<T extends State>(target: T): State.Values<T> {
 
   return Object.freeze(values);
 }
-
-type Proxy<T = any> = (key: string | number, value: T) => T;
-
-const OBSERVER = new WeakMap<State, Proxy>();
-
-function follow(from: State, key: string | number, value?: any) {
-  const observe = OBSERVER.get(from);
-  return observe ? observe(key, value) : value;
-}
-
 function access(subject: State, property: string, required?: boolean) {
   const state = STATE.get(subject)!;
 
@@ -720,4 +693,4 @@ function uid() {
     .toUpperCase();
 }
 
-export { event, METHOD, State, PARENT, STATE, uid, access, update, follow };
+export { event, METHOD, State, PARENT, STATE, uid, access, update };
